ball trajectory working
