+++ com.barrycodes.jsgame01/src/main/java/jsgame01/services/GameInstanceService.java
@@ -19,4 +19,8 @@ public interface GameInstanceService {
 
     public void deleteAllGameInstances();
 
+    public Iterable<GameInstance> getTop10Scores();
+
+    public Iterable<GameInstance> getTop10ScoresToday();
+
 }
