+++ com.barrycodes.jsgame01/src/main/webapp/static/js/game/settings.js
@@ -1,0 +1,24 @@
+var game;
+var slider;
+var sliderTimer;
+var keyLeft = 37;
+var keyRight = 39;
+var keyEscape = 13;
+var active = false;
+var movingRight = false;
+var movingLeft = false;
+var sliderPosition = 20.6;
+var minSliderPosition = 2;
+var maxSliderPosition = 39.2;
+var frameDelay = 10;
+var sliderDistancePerFrame = 0.4;
+var readyDelay = 2000;
+var goDelay = 350;
+var ball;
+var ballDistancePerFrame = 0.1;
+var ballTimer;
+var ballX = 19;
+var ballY = -20;
+var ballAngle = 350;
+var ballDeltaX;
+var ballDeltaY;
